Refactor account propagation + infinite loop
Remove some println
